Move `AddPaymentMethod` to its own file
